Refactor code, moving templating logic to TemplateService
