+++ packages/discovery/src/discovery/config/ConfigReader.ts
@@ -3,16 +3,16 @@ import path, { posix } from 'path'
 import { assert } from '@l2beat/backend-tools'
 import { DiscoveryOutput } from '@l2beat/discovery-types'
 import { readFile } from 'fs/promises'
-import { ParseError, parse } from 'jsonc-parser'
 
 import { stripAnsiEscapeCodes } from '@l2beat/shared-pure'
 import chalk from 'chalk'
 import { ZodError } from 'zod'
 import { fileExistsCaseSensitive } from '../../utils/fsLayer'
 import { TemplateService } from '../analysis/TemplateService'
+import { readJsonc } from '../utils/readJsonc'
 import { DiscoveryConfig } from './DiscoveryConfig'
-import { ContractMeta, DiscoveryMeta } from './DiscoveryMeta'
-import { DiscoveryContract, RawDiscoveryConfig } from './RawDiscoveryConfig'
+import { DiscoveryMeta } from './DiscoveryMeta'
+import { RawDiscoveryConfig } from './RawDiscoveryConfig'
 
 export const TEMPLATES_PATH = path.join('discovery', '_templates')
 
@@ -33,9 +33,7 @@ export class ConfigReader {
       'Chain not found in project, check if case matches',
     )
 
-    const contents = await this.readJsonc(
-      `discovery/${name}/${chain}/config.jsonc`,
-    )
+    const contents = await readJsonc(`discovery/${name}/${chain}/config.jsonc`)
     const rawConfig = RawDiscoveryConfig.safeParse(contents)
     if (!rawConfig.success) {
       const message = formatZodParsingError(rawConfig.error, 'config.jsonc')
@@ -44,42 +42,12 @@ export class ConfigReader {
       throw new Error(`Cannot parse file ${name}/${chain}/config.jsonc`)
     }
 
-    await this.inlineTemplates(rawConfig.data)
+    await this.templateService.inlineTemplates(rawConfig.data)
     const config = new DiscoveryConfig(rawConfig.data)
 
     assert(config.chain === chain, 'Chain mismatch in config.jsonc')
 
     return config
-  }
-
-  async readJsonc(path: string): Promise<JSON> {
-    const contents = await readFile(path, 'utf-8')
-    const errors: ParseError[] = []
-    const parsed = parse(contents, errors, {
-      allowTrailingComma: true,
-    }) as JSON
-    if (errors.length !== 0) {
-      throw new Error(`Cannot parse file ${path}`)
-    }
-    return parsed
-  }
-
-  async inlineTemplates(rawConfig: RawDiscoveryConfig): Promise<void> {
-    if (rawConfig.overrides === undefined) {
-      return
-    }
-    for (const [name, contract] of Object.entries(rawConfig.overrides)) {
-      if (contract.extends !== undefined) {
-        const templateJson = await this.readJsonc(
-          path.join(TEMPLATES_PATH, contract.extends, 'template.jsonc'),
-        )
-        const updatedContract = DiscoveryContract.parse({
-          ...templateJson,
-          ...contract,
-        })
-        rawConfig.overrides[name] = updatedContract
-      }
-    }
   }
 
   async readMeta(
@@ -108,37 +76,10 @@ export class ConfigReader {
 
     const meta = DiscoveryMeta.parse(JSON.parse(contents))
     if (!skipTemplates) {
-      this.inlineMetaTemplates(meta)
+      this.templateService.inlineMetaTemplates(meta)
       meta._templatesWereInlined = true
     }
     return meta
-  }
-
-  inlineMetaTemplates(rawMeta: DiscoveryMeta): void {
-    const expandedContracts: ContractMeta[] = []
-    for (const contract of rawMeta.contracts) {
-      if (contract.extends !== undefined) {
-        const template = this.templateService.readContractMetaTemplate(
-          contract.extends,
-        )
-        const expandedContract = ContractMeta.parse({
-          ...template,
-          ...contract,
-        })
-        // merge each value separately to not overwrite the whole object
-        expandedContract.values = template.values ?? {}
-        for (const [key, value] of Object.entries(contract.values ?? {})) {
-          expandedContract.values[key] = {
-            ...expandedContract.values[key],
-            ...value,
-          }
-        }
-        expandedContracts.push(expandedContract)
-      } else {
-        expandedContracts.push(contract)
-      }
-    }
-    rawMeta.contracts = expandedContracts
   }
 
   async readDiscovery(name: string, chain: string): Promise<DiscoveryOutput> {
